refactor: 사용하지 않는 Mappging 제거 및 ReadStatus와 UserStatus 위치 변경
