--- src/main/java/com/sprint/mission/discodeit/config/MappingConfig.java
@@ -1,8 +1,0 @@
-package com.sprint.mission.discodeit.config;
-
-import org.mapstruct.MapperConfig;
-
-@MapperConfig(componentModel = "spring")
-public class MappingConfig {
-
-}
