+++ src/main/java/com/sprint/mission/discodeit/core/status/controller/ReadStatusController.java
@@ -3,9 +3,9 @@ package com.sprint.mission.discodeit.core.status.controller;
 import com.sprint.mission.discodeit.core.status.controller.dto.ReadStatusCreateRequest;
 import com.sprint.mission.discodeit.core.status.controller.dto.ReadStatusUpdateRequest;
 import com.sprint.mission.discodeit.core.status.usecase.read.ReadStatusService;
-import com.sprint.mission.discodeit.core.status.usecase.read.dto.ReadStatusCreateCommand;
-import com.sprint.mission.discodeit.core.status.usecase.read.dto.ReadStatusDto;
-import com.sprint.mission.discodeit.core.status.usecase.read.dto.ReadStatusUpdateCommand;
+import com.sprint.mission.discodeit.core.status.usecase.dto.ReadStatusCreateCommand;
+import com.sprint.mission.discodeit.core.status.usecase.dto.ReadStatusDto;
+import com.sprint.mission.discodeit.core.status.usecase.dto.ReadStatusUpdateCommand;
 import com.sprint.mission.discodeit.swagger.ReadStatusApi;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
