+++ src/main/java/com/sprint/mission/discodeit/core/status/usecase/dto/ReadStatusUpdateCommand.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.core.status.usecase.read.dto;
+package com.sprint.mission.discodeit.core.status.usecase.dto;
 
 import java.time.Instant;
 import java.util.UUID;
