+++ src/main/java/com/sprint/mission/discodeit/core/status/usecase/read/CreateReadStatusUseCase.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.core.status.usecase.read;
 
-import com.sprint.mission.discodeit.core.status.usecase.read.dto.ReadStatusCreateCommand;
-import com.sprint.mission.discodeit.core.status.usecase.read.dto.ReadStatusDto;
+import com.sprint.mission.discodeit.core.status.usecase.dto.ReadStatusCreateCommand;
+import com.sprint.mission.discodeit.core.status.usecase.dto.ReadStatusDto;
 
 public interface CreateReadStatusUseCase {
