+++ src/main/java/com/sprint/mission/discodeit/core/status/usecase/read/UpdateReadStatusUseCase.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.core.status.usecase.read;
 
-import com.sprint.mission.discodeit.core.status.usecase.read.dto.ReadStatusDto;
-import com.sprint.mission.discodeit.core.status.usecase.read.dto.ReadStatusUpdateCommand;
+import com.sprint.mission.discodeit.core.status.usecase.dto.ReadStatusDto;
+import com.sprint.mission.discodeit.core.status.usecase.dto.ReadStatusUpdateCommand;
 
 public interface UpdateReadStatusUseCase {
