+++ src/main/java/com/sprint/mission/discodeit/core/status/usecase/user/CreateUserStatusUseCase.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.core.status.usecase.user;
 
 import com.sprint.mission.discodeit.core.status.entity.UserStatus;
-import com.sprint.mission.discodeit.core.status.usecase.user.dto.UserStatusCreateCommand;
+import com.sprint.mission.discodeit.core.status.usecase.dto.UserStatusCreateCommand;
 
 public interface CreateUserStatusUseCase {
