+++ src/main/java/com/sprint/mission/discodeit/core/status/usecase/user/UpdateUserStatusUseCase.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.core.status.usecase.user;
 
 import com.sprint.mission.discodeit.core.status.entity.UserStatus;
-import com.sprint.mission.discodeit.core.status.usecase.user.dto.UserStatusUpdateCommand;
+import com.sprint.mission.discodeit.core.status.usecase.dto.UserStatusUpdateCommand;
 
 public interface UpdateUserStatusUseCase {
