+++ src/main/java/com/sprint/mission/discodeit/core/user/controller/UserController.java
@@ -2,8 +2,8 @@ package com.sprint.mission.discodeit.core.user.controller;
 
 
 import com.sprint.mission.discodeit.core.content.usecase.dto.BinaryContentCreateCommand;
-import com.sprint.mission.discodeit.core.status.usecase.user.dto.UserStatusOnlineCommand;
-import com.sprint.mission.discodeit.core.status.usecase.user.dto.UserStatusDto;
+import com.sprint.mission.discodeit.core.status.usecase.dto.UserStatusOnlineCommand;
+import com.sprint.mission.discodeit.core.status.usecase.dto.UserStatusDto;
 import com.sprint.mission.discodeit.core.user.controller.dto.UserCreateRequest;
 import com.sprint.mission.discodeit.core.user.controller.dto.UserStatusRequest;
 import com.sprint.mission.discodeit.core.user.controller.dto.UserUpdateRequest;
