+++ src/main/java/com/sprint/mission/discodeit/core/user/usecase/BasicUserService.java
@@ -6,9 +6,9 @@ import com.sprint.mission.discodeit.core.content.usecase.BinaryContentService;
 import com.sprint.mission.discodeit.core.content.usecase.dto.BinaryContentCreateCommand;
 import com.sprint.mission.discodeit.core.status.entity.UserStatus;
 import com.sprint.mission.discodeit.core.status.usecase.user.UserStatusService;
-import com.sprint.mission.discodeit.core.status.usecase.user.dto.UserStatusCreateCommand;
-import com.sprint.mission.discodeit.core.status.usecase.user.dto.UserStatusOnlineCommand;
-import com.sprint.mission.discodeit.core.status.usecase.user.dto.UserStatusDto;
+import com.sprint.mission.discodeit.core.status.usecase.dto.UserStatusCreateCommand;
+import com.sprint.mission.discodeit.core.status.usecase.dto.UserStatusOnlineCommand;
+import com.sprint.mission.discodeit.core.status.usecase.dto.UserStatusDto;
 import com.sprint.mission.discodeit.core.user.entity.User;
 import com.sprint.mission.discodeit.core.user.exception.UserAlreadyExistsException;
 import com.sprint.mission.discodeit.core.user.exception.UserLoginFailedException;
