+++ src/main/java/com/sprint/mission/discodeit/core/user/usecase/OnlineUserUseCase.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.core.user.usecase;
 
-import com.sprint.mission.discodeit.core.status.usecase.user.dto.UserStatusOnlineCommand;
-import com.sprint.mission.discodeit.core.status.usecase.user.dto.UserStatusDto;
+import com.sprint.mission.discodeit.core.status.usecase.dto.UserStatusOnlineCommand;
+import com.sprint.mission.discodeit.core.status.usecase.dto.UserStatusDto;
 
 public interface OnlineUserUseCase {
