+++ src/test/java/com/sprint/mission/discodeit/service/UserServiceUnitTest.java
@@ -19,8 +19,8 @@ import com.sprint.mission.discodeit.core.content.usecase.BinaryContentService;
 import com.sprint.mission.discodeit.core.content.usecase.dto.BinaryContentCreateCommand;
 import com.sprint.mission.discodeit.core.status.entity.UserStatus;
 import com.sprint.mission.discodeit.core.status.usecase.user.UserStatusService;
-import com.sprint.mission.discodeit.core.status.usecase.user.dto.UserStatusDto;
-import com.sprint.mission.discodeit.core.status.usecase.user.dto.UserStatusOnlineCommand;
+import com.sprint.mission.discodeit.core.status.usecase.dto.UserStatusDto;
+import com.sprint.mission.discodeit.core.status.usecase.dto.UserStatusOnlineCommand;
 import com.sprint.mission.discodeit.core.user.entity.User;
 import com.sprint.mission.discodeit.core.user.exception.UserAlreadyExistsException;
 import com.sprint.mission.discodeit.core.user.exception.UserInvalidRequestException;
